+++ test-package-compiler/src/generated_tests.rs
@@ -58,6 +58,17 @@ fn duplicate_module_test_dev() {
 
 #[rustfmt::skip]
 #[test]
+fn empty_module_warning() {
+    let output = crate::prepare("./cases/empty_module_warning");
+    insta::assert_snapshot!(
+        "empty_module_warning",
+        output,
+        "./cases/empty_module_warning",
+    );
+}
+
+#[rustfmt::skip]
+#[test]
 fn erlang_app_generation() {
     let output = crate::prepare("./cases/erlang_app_generation");
     insta::assert_snapshot!(