+++ CHANGELOG.md
@@ -361,6 +361,10 @@
   has been improved to avoid performing needless checks.
   ([Giacomo Cavalieri](https://github.com/giacomocavalieri))
 
+- The compiler now emits a warning when a module contains no public definitions
+  and prevents publishing packages with empty modules to Hex.
+  ([Vitor Souza](https://github.com/vit0rr))
+
 ### Build tool
 
 - `gleam update`, `gleam deps update`, and `gleam deps download` will now print